feat(sound): driver-agnostic ALSA core — ALSA formats, card identity, control elements
The sound core was virtio-shaped: capability masks were indexed by
virtio-snd's private format enum, hw_params took a virtio rate code,
card and PCM names were the string "virtio-snd", and the control node
could never expose a mixer element. A second sound card cannot express
its capabilities through that boundary at all — HDA's S24/S32 formats
have no virtio code — and a second PCM implementation beside it would be
a split source of truth.

Formats crossing the driver boundary are now SNDRV_PCM_FORMAT_* values
and rates are Hz; sound::format owns the one rate table (it had three
copies) and the container-width math. Card identity, transfer-path
limits and SNDRV_PCM_INFO_* capability bits come from the card's ops, so
PAUSE is refused with ENOSYS on a card that cannot pause and honoured on
one that can. sound::elem is a real control-element registry with
list/info/read/write/TLV and a bounded per-card event queue that a
control fd can read and poll. virtio-snd keeps its own encoding in
drv-virtio-snd::fmt and translates at the boundary.

+++ crates/kernel/sound/src/oss_state.rs
@@ -4,7 +4,7 @@ use sync::{Spinlock, TaskList as L};
 pub(crate) struct Oss {
     pub owner: crate::SoundOwnerKey,
     pub rate: u8,
-    pub format: u8,
+    pub format: u32,
     pub channels: u8,
     pub subdivision: u8,
     pub fragshift: u8,
